feat: 敏感信息存储支持国密 #2055
升级SDK版本，优化代码结构
