+++ src/backend/commons/common-crypto/src/main/java/com.tencent.bk.job.common/crypto/RSAEncryptor.java
@@ -22,16 +22,39 @@
  * IN THE SOFTWARE.
  */
 
-package com.tencent.bk.job.common.encrypt;
+package com.tencent.bk.job.common.crypto;
 
-/**
- * 加密类型枚举值
- */
-public enum CryptoTypeEnum {
+import com.tencent.bk.job.common.util.crypto.RSAUtils;
 
-    // 经典密码算法（RSA、AES等）
-    CLASSIC,
-    // 国家商用密码算法（SM2、SM4等）
-    SHANGMI
+import java.io.File;
+import java.io.IOException;
+import java.security.GeneralSecurityException;
+import java.security.PublicKey;
 
+public class RSAEncryptor implements AsymmetricEncryptor {
+    private PublicKey publicKey;
+
+    public RSAEncryptor(File rsaPubPermFile) throws IOException, GeneralSecurityException {
+        publicKey = RSAUtils.getPublicKey(rsaPubPermFile);
+    }
+
+    public RSAEncryptor(String rsaPublicKeyBase64) throws IOException, GeneralSecurityException {
+        publicKey = RSAUtils.getPublicKey(rsaPublicKeyBase64);
+    }
+
+    public String encrypt(String rawText) {
+        try {
+            return RSAUtils.encrypt(rawText, publicKey);
+        } catch (IOException | GeneralSecurityException e) {
+            return null;
+        }
+    }
+
+    public boolean verify(String message, String signature) {
+        try {
+            return RSAUtils.verify(publicKey, message, signature);
+        } catch (Exception e) {
+            return false;
+        }
+    }
 }
