+++ src/backend/commons/common-crypto/src/main/java/com.tencent.bk.job.common/crypto/scenario/SensitiveParamCryptoService.java
@@ -22,13 +22,14 @@
  * IN THE SOFTWARE.
  */
 
-package com.tencent.bk.job.common.encrypt.scenario;
+package com.tencent.bk.job.common.crypto.scenario;
 
-import com.cronutils.utils.StringUtils;
-import com.tencent.bk.job.common.encrypt.CryptoScenarioEnum;
-import com.tencent.bk.job.common.encrypt.SymmetricCryptoService;
+import com.tencent.bk.job.common.crypto.CryptoScenarioEnum;
+import com.tencent.bk.job.common.crypto.SymmetricCryptoService;
 import com.tencent.bk.sdk.crypto.cryptor.consts.CryptorNames;
+import com.tencent.bk.sdk.crypto.util.CryptorMetaUtil;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -50,7 +51,7 @@ public class SensitiveParamCryptoService {
         if (!secureParam || StringUtils.isEmpty(cipher)) {
             return CryptorNames.NONE;
         }
-        String algorithm = symmetricCryptoService.getAlgorithmFromCipher(cipher);
+        String algorithm = CryptorMetaUtil.getCryptorNameFromCipher(cipher);
         if (algorithm != null) {
             return algorithm;
         }
