+++ src/backend/commons/common-crypto/src/main/resources/META-INF/services/com.tencent.bk.sdk.crypto.cryptor.SymmetricCryptor
@@ -1,0 +1,1 @@
+com.tencent.bk.job.common.crypto.AESCryptor
