--- src/backend/commons/common/src/main/resources/META-INF/services/com.tencent.bk.sdk.crypto.cryptor.SymmetricCryptor
@@ -1,1 +1,0 @@
-com.tencent.bk.job.common.encrypt.AESCryptor
