+++ src/backend/job-backup/service-job-backup/src/main/java/com/tencent/bk/job/backup/crypto/BackupFileCryptoService.java
@@ -22,52 +22,60 @@
  * IN THE SOFTWARE.
  */
 
-package com.tencent.bk.job.common.encrypt.scenario;
+package com.tencent.bk.job.backup.crypto;
 
-import com.tencent.bk.job.common.encrypt.CryptoScenarioEnum;
-import com.tencent.bk.job.common.encrypt.JobCryptorNames;
-import com.tencent.bk.job.common.encrypt.SymmetricCryptoService;
-import com.tencent.bk.sdk.crypto.cryptor.consts.CryptorNames;
+import com.tencent.bk.job.common.crypto.CryptoScenarioEnum;
+import com.tencent.bk.job.common.crypto.JobCryptorNames;
+import com.tencent.bk.job.common.crypto.SymmetricCryptoService;
+import com.tencent.bk.job.common.exception.CryptoException;
+import com.tencent.bk.sdk.crypto.util.CryptorMetaUtil;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.io.BufferedInputStream;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
+
 /**
- * DB账号密码相关加解密服务
+ * 导入导出文件相关加解密服务
  */
 @Slf4j
 @Service
-public class CredentialCryptoService {
+public class BackupFileCryptoService {
 
     private final SymmetricCryptoService symmetricCryptoService;
 
     @Autowired
-    public CredentialCryptoService(SymmetricCryptoService symmetricCryptoService) {
+    public BackupFileCryptoService(SymmetricCryptoService symmetricCryptoService) {
         this.symmetricCryptoService = symmetricCryptoService;
     }
 
-    public String getCredentialEncryptAlgorithmByCipher(String cipher) {
-        if (StringUtils.isEmpty(cipher)) {
-            return CryptorNames.NONE;
+    public void encryptBackupFile(String password, File inFile, File outFile) {
+        try (FileInputStream in = new FileInputStream(inFile); FileOutputStream out = new FileOutputStream(outFile)) {
+            symmetricCryptoService.encrypt(password, in, out, CryptoScenarioEnum.BACKUP_FILE);
+        } catch (Exception e) {
+            throw new CryptoException("Fail to encrypt backupFile", e);
         }
-        String algorithm = symmetricCryptoService.getAlgorithmFromCipher(cipher);
-        if (algorithm != null) {
-            return algorithm;
-        }
-        return JobCryptorNames.AES_CBC;
     }
 
-    public String encryptCredential(String credentialValue) {
-        return symmetricCryptoService.encryptToBase64Str(credentialValue, CryptoScenarioEnum.CREDENTIAL);
-    }
-
-    public String decryptCredential(String encryptedCredential) {
-        String algorithm = getCredentialEncryptAlgorithmByCipher(encryptedCredential);
-        if (StringUtils.isBlank(algorithm)) {
-            return encryptedCredential;
+    public void decryptBackupFile(String password, File inFile, File outFile) {
+        try {
+            FileInputStream in = new FileInputStream(inFile);
+            BufferedInputStream bis = new BufferedInputStream(in);
+            String algorithm = CryptorMetaUtil.getCryptorNameFromCipherStream(bis);
+            if (StringUtils.isBlank(algorithm)) {
+                algorithm = JobCryptorNames.AES_CBC;
+            }
+            try (FileOutputStream out = new FileOutputStream(outFile)) {
+                log.debug("Use {} to decryptBackupFile", algorithm);
+                symmetricCryptoService.decrypt(password, bis, out, algorithm);
+            }
+        } catch (Exception e) {
+            throw new CryptoException("Fail to decrypt backupFile", e);
         }
-        return symmetricCryptoService.decrypt(encryptedCredential, algorithm);
     }
 
 }
