+++ src/backend/job-crontab/service-job-crontab/build.gradle
@@ -24,6 +24,7 @@
 
 apply plugin: 'io.spring.dependency-management'
 dependencies {
+    api project(":commons:common-crypto")
     api project(":commons:common-redis")
     api project(":commons:common-web")
     api project(':commons:common-security')
