+++ src/backend/job-manage/boot-job-manage/src/main/java/com/tencent/bk/job/manage/JobManageBootApplication.java
@@ -25,7 +25,7 @@
 package com.tencent.bk.job.manage;
 
 import com.tencent.bk.job.common.config.FeatureToggleConfig;
-import com.tencent.bk.job.common.encrypt.EncryptConfig;
+import com.tencent.bk.job.common.crypto.EncryptConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.data.redis.RedisAutoConfiguration;
