+++ src/backend/job-manage/service-job-manage/src/main/java/com/tencent/bk/job/manage/dao/template/impl/TaskTemplateScriptStepDAOImpl.java
@@ -24,7 +24,7 @@
 
 package com.tencent.bk.job.manage.dao.template.impl;
 
-import com.tencent.bk.job.common.encrypt.scenario.SensitiveParamCryptoService;
+import com.tencent.bk.job.common.crypto.scenario.SensitiveParamCryptoService;
 import com.tencent.bk.job.manage.common.consts.script.ScriptTypeEnum;
 import com.tencent.bk.job.manage.common.consts.task.TaskScriptSourceEnum;
 import com.tencent.bk.job.manage.common.util.JooqDataTypeUtil;
