+++ src/backend/settings.gradle
@@ -39,6 +39,7 @@ include 'commons:common-redis'
 include 'commons:common-security'
 include 'commons:common-discovery'
 include 'commons:common-consul'
+include 'commons:common-crypto'
 include 'commons:common-k8s'
 include 'commons:common-statistics'
 include 'commons:common-service'
